+++ docs/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.19.6](https://github.com/samchen08/vtj.pro/compare/@vtj/docs@0.19.5...@vtj/docs@0.19.6) (2026-08-14)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
+## [0.19.5](https://github.com/samchen08/vtj.pro/compare/@vtj/docs@0.19.4...@vtj/docs@0.19.5) (2026-08-11)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
+## [0.19.4](https://github.com/samchen08/vtj.pro/compare/@vtj/docs@0.19.3...@vtj/docs@0.19.4) (2026-08-09)
+
+**Note:** Version bump only for package @vtj/docs
+
+
+
+
+
 ## [0.19.3](https://github.com/samchen08/vtj.pro/compare/@vtj/docs@0.19.2...@vtj/docs@0.19.3) (2026-08-09)
 
 **Note:** Version bump only for package @vtj/docs