+++ packages/renderer/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.19.6](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.19.5...@vtj/renderer@0.19.6) (2026-08-14)
+
+**Note:** Version bump only for package @vtj/renderer
+
+
+
+
+
+## [0.19.5](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.19.4...@vtj/renderer@0.19.5) (2026-08-11)
+
+**Note:** Version bump only for package @vtj/renderer
+
+
+
+
+
+## [0.19.4](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.19.3...@vtj/renderer@0.19.4) (2026-08-09)
+
+**Note:** Version bump only for package @vtj/renderer
+
+
+
+
+
 ## [0.19.3](https://gitee.com/newgateway/vtj/compare/@vtj/renderer@0.19.2...@vtj/renderer@0.19.3) (2026-08-09)
 
 **Note:** Version bump only for package @vtj/renderer